+++ pkgs/jni/lib/src/jarray.dart
@@ -468,6 +468,13 @@ final class JByteArrayType extends JObjType<JByteArray> {
   }
 }
 
+/// A fixed-length array of Java [`Byte`](https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Byte.html).
+///
+/// Integers stored in the list are truncated to their low eight bits,
+/// interpreted as a signed 8-bit two's complement integer with values in the
+/// range -128 to +127.
+///
+/// Java equivalent of [Int8List].
 class JByteArray extends JObject with Iterable<int> {
   @internal
   @override
@@ -484,6 +491,16 @@ class JByteArray extends JObject with Iterable<int> {
   JByteArray.fromReference(super.reference)
       : $type = type,
         super.fromReference();
+
+  /// Creates a [JByteArray] containing all `elements`.
+  ///
+  /// The [Iterator] of elements provides the order of the elements.
+  ///
+  /// Elements outside of the range -128 to +127 are truncated to their low
+  /// eight bits and interpreted as signed 8-bit two's complement integers.
+  factory JByteArray.from(Iterable<int> elements) {
+    return JByteArray(elements.length)..setRange(0, elements.length, elements);
+  }
 
   /// Creates a [JByteArray] of the given [length].
   ///