+++ pkgs/jni/CHANGELOG.md
@@ -12,6 +12,8 @@
 - Added `JObject.isA`, which checks whether a `JObject` is a instance of a
   java class.
 - Do not require JWT when building for desktop.
+- Added `JByteArray.from`, which allows a `JByteArray` to be constructed
+  from an `Iterable<int>`.
 
 ## 0.13.0